refactor(#60): cleanup current state of the dashboard
Restruture current structure of the dashbaord by using
subdirectories and updating file-names for more
consistency with the other parts of the project.
