+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__
 docs_env
 ainari_env
 target
+.vite
 
 testing/go_cli_api/ainarictl
 testing/ansible_deploy/.vagrant/
