+++ src/dashboard/app/src/components/admin/project.vue
@@ -5,7 +5,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 
-//     http://www.apache.org/licenses/LICENSE-2.0
+//         http://www.apache.org/licenses/LICENSE-2.0
 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,37 +17,35 @@
 <template>
     <div class="overview">
         <div class="card">
-            <div class="card-label">Users</div>
+            <div class="card-label">Project</div>
             <div class="card-content">
                 <!-- Add button -->
                 <button class="add-button" @click="openAddModal">+</button>
 
-                <table v-if="users.length > 0">
+                <table v-if="projects.length > 0">
                     <thead>
                         <tr>
                             <th>ID</th>
-                            <th>Username</th>
-                            <th>Is Admin</th>
                             <th>Actions</th>
                         </tr>
                     </thead>
                     <tbody>
-                        <tr v-for="user in users" :key="user.id">
-                            <td>{{ user.id }}</td>
-                            <td>{{ user.name }}</td>
-                            <td>{{ user.is_admin }}</td>
+                        <tr v-for="project in projects" :key="project.id">
+                            <td>{{ project.id }}</td>
                             <td>
                                 <!-- Dropdown menu -->
                                 <div
                                     class="table-dropdown"
-                                    @click.stop="toggleDropdown(user.id)"
+                                    @click.stop="toggleDropdown(project.id)"
                                 >
                                     ⋮
                                     <div
-                                        v-if="openDropdown === user.id"
+                                        v-if="openDropdown === project.id"
                                         class="table-dropdown-menu"
                                     >
-                                        <button @click="openDeleteModal(user)">
+                                        <button
+                                            @click="openDeleteModal(project)"
+                                        >
                                             Delete
                                         </button>
                                     </div>
@@ -57,54 +55,34 @@
                     </tbody>
                 </table>
 
-                <p v-else>No users found</p>
+                <p v-else>No projects found</p>
             </div>
         </div>
 
-        <!-- Add User Modal -->
+        <!-- Add Project Modal -->
         <div
             v-if="showAddModal"
             class="modal-overlay"
             @click.self="cancelAddModal"
         >
-            <div class="modal user-create-modal">
+            <div class="modal project-create-modal">
                 <!-- Modal topbar -->
                 <div class="modal-topbar">
-                    <span>Create user</span>
+                    <span>Create project</span>
                 </div>
                 <div class="modal-content">
                     <input
-                        v-model="newUser.userid"
+                        v-model="newProject.projectId"
                         type="text"
-                        placeholder="User-ID"
+                        placeholder="Project-ID"
                         required
                     />
                     <input
-                        v-model="newUser.username"
+                        v-model="newProject.projectName"
                         type="text"
-                        placeholder="User-Name"
+                        placeholder="Project-Name"
                         required
                     />
-                    <input
-                        v-model="newUser.password"
-                        type="password"
-                        placeholder="Password"
-                        required
-                    />
-                    <input
-                        v-model="newUser.confirmPassword"
-                        type="password"
-                        placeholder="Confirm password"
-                        required
-                    />
-                    <label class="checkbox-label">
-                        <input type="checkbox" v-model="newUser.isAdmin" />
-                        Is Admin
-                    </label>
-
-                    <p v-if="passwordError" class="error-msg">
-                        {{ passwordError }}
-                    </p>
                 </div>
 
                 <div class="modal-bottombar">
@@ -126,13 +104,13 @@
             class="modal-overlay"
             @click.self="cancelDeleteModal"
         >
-            <div class="modal user-delete-modal">
+            <div class="modal project-delete-modal">
                 <div class="modal-topbar">
-                    <span>Delete user</span>
+                    <span>Delete project</span>
                 </div>
                 <div class="modal-content">
                     <p>Are you sure you want to delete?</p>
-                    <strong>User: {{ userToDelete?.id }}</strong>
+                    <strong>Project: {{ projectToDelete?.id }}</strong>
                 </div>
 
                 <div class="modal-bottombar">
@@ -152,39 +130,36 @@
 
 <script setup lang="ts">
 import { ref, onMounted, onBeforeUnmount, inject } from "vue";
-import api from "../api";
+import api from "../../api";
 
-const users = ref<{ id: number; username: string }[]>([]);
+const projects = ref<{ id: string; projectName: string; email: string }[]>([]);
 const showAddModal = ref(false);
 const showDeleteModal = ref(false);
-const openDropdown = ref<number | null>(null);
-const newUser = ref({
-    userid: "",
-    username: "",
-    password: "",
-    confirmPassword: "",
-    isAdmin: false,
+const openDropdown = ref<string | null>(null);
+const newProject = ref({
+    projectId: "",
+    projectName: "",
 });
 const passwordError = ref("");
-const userToDelete = ref<{ id: number; username: string } | null>(null);
+const projectToDelete = ref<{ id: string; projectName: string } | null>(null);
 const icons = inject<{ acceptIcon: string; cancelIcon: string }>("icons")!;
 
-async function fetchUsers() {
+async function fetchProjects() {
     try {
         const token = localStorage.getItem("jwtToken");
-        const response = await api.torii_api.get("/v1alpha/user", {
+        const response = await api.torii_api.get("/v1alpha/project", {
             headers: { Authorization: `Bearer ${token}` },
         });
-        users.value = response.data.users;
+        projects.value = response.data.projects;
     } catch (err) {
-        console.error("Failed to load users", err);
+        console.error("Failed to load projects", err);
     }
 }
 
 //=============================================================================
 // Dropdown in table
 //=============================================================================
-function toggleDropdown(id: number) {
+function toggleDropdown(id: string) {
     openDropdown.value = openDropdown.value === id ? null : id;
 }
 
@@ -202,71 +177,74 @@ function handleClickOutside(event: MouseEvent) {
 }
 
 //=============================================================================
-// Add user modal
+// Add project modal
 //=============================================================================
 function openAddModal() {
     showAddModal.value = true;
 }
 function cancelAddModal() {
     showAddModal.value = false;
-    newUser.value.userid = "";
-    newUser.value.username = "";
-    newUser.value.password = "";
-    newUser.value.confirmPassword = "";
-    newUser.value.isAdmin = false;
+    newProject.value.projectId = "";
+    newProject.value.projectName = "";
+    newProject.value.password = "";
+    newProject.value.confirmPassword = "";
+    newProject.value.isAdmin = false;
     passwordError.value = "";
 }
 
 async function acceptAddModal() {
-    if (newUser.value.password !== newUser.value.confirmPassword) {
+    if (newProject.value.password !== newProject.value.confirmPassword) {
         passwordError.value = "Passwords do not match!";
         return;
     }
     try {
         const token = localStorage.getItem("jwtToken");
         await api.torii_api.post(
-            "/v1alpha/user",
+            "/v1alpha/project",
             {
-                id: newUser.value.userid,
-                name: newUser.value.username,
-                passphrase: newUser.value.password,
-                is_admin: newUser.value.isAdmin,
+                id: newProject.value.projectId,
+                name: newProject.value.projectName,
+                passphrase: newProject.value.password,
+                is_admin: newProject.value.isAdmin,
             },
             {
                 headers: { Authorization: `Bearer ${token}` },
             },
         );
-        await fetchUsers();
+        await fetchProjects();
         cancelAddModal();
     } catch (err) {
         passwordError.value = err;
-        console.error("Failed to create user", err);
+        console.error("Failed to create project", err);
     }
 }
 
 //=============================================================================
 // Delete modal
 //=============================================================================
-function openDeleteModal(user: { id: number; username: string }) {
-    userToDelete.value = user;
+function openDeleteModal(project: { id: string; projectName: string }) {
+    projectToDelete.value = project;
     showDeleteModal.value = true;
     openDropdown.value = null;
 }
 function cancelDeleteModal() {
     showDeleteModal.value = false;
-    userToDelete.value = null;
+    projectToDelete.value = null;
     openDropdown.value = null; // close any open action dropdown
 }
 async function acceptDeleteModal() {
-    if (!userToDelete.value) return;
+    if (!projectToDelete.value) return;
     try {
         const token = localStorage.getItem("jwtToken");
-        await api.torii_api.delete(`/v1alpha/user/${userToDelete.value.id}`, {
-            headers: { Authorization: `Bearer ${token}` },
-        });
-        await fetchUsers();
+        await api.torii_api.delete(
+            `/v1alpha/project/${projectToDelete.value.id}`,
+            {
+                headers: { Authorization: `Bearer ${token}` },
+            },
+        );
+        await fetchProjects();
     } catch (err) {
-        console.error("Failed to delete user", err);
+        console.error("Failed to delete project", err);
     } finally {
         cancelDeleteModal();
     }
@@ -275,7 +253,7 @@ async function acceptDeleteModal() {
 //=============================================================================
 // Listener
 //=============================================================================
-onMounted(fetchUsers);
+onMounted(fetchProjects);
 
 onMounted(() => {
     window.addEventListener("click", handleClickOutside);
@@ -287,12 +265,12 @@ onBeforeUnmount(() => {
 </script>
 
 <style scoped>
-.user-create-modal {
-    height: 28rem;
+.project-create-modal {
+    height: 18rem;
     width: 20rem;
 }
 
-.user-delete-modal {
+.project-delete-modal {
     height: 16rem;
     width: 20rem;
 }
