+++ src/dashboard/app/src/styles/base.css
@@ -143,4 +143,3 @@ label {
     text-align: left;
     font-size: 0.9rem;
 }
-
