+++ src/dashboard/app/src/styles/tab.css
@@ -48,4 +48,3 @@ limitations under the License.
     background-color: var(--color-text-light);
     color: black;
 }
-
